cyber_reasoning_system/patch_generation.py
## patch_generation.py
class PatchGeneration:
    def __init__(self):
        self.patches = []

    def generate_patch(self, vulnerability: str):
        """
        Generate patches for the given vulnerability.
        The actual implementation of this function will depend on the specific
        vulnerability and the patching technique used. For the purpose of this
        example, we will simply return a dummy patch.
        """
        patch = {
            'vulnerability': vulnerability,
            'patch': 'Dummy patch for ' + vulnerability
        }
        self.patches.append(patch)
        return self.patches
