tests/test_patch_generation.py
## test_patch_generation.py

import unittest
from cyber_reasoning_system.patch_generation import PatchGeneration

class TestPatchGeneration(unittest.TestCase):
    """
    Test cases for PatchGeneration class
    """

    ## Test Initialization
    def test_initialization(self):
        """
        Test that the PatchGeneration object is initialized correctly
        """
        patch_gen = PatchGeneration()
        self.assertEqual(patch_gen.patches, [])

    ## Test generate_patch method
    def test_generate_patch(self):
        """
        Test that the generate_patch method works correctly
        """
        patch_gen = PatchGeneration()
        patch = patch_gen.generate_patch('vulnerability1')
        self.assertEqual(len(patch), 1)
        self.assertEqual(patch[0]['vulnerability'], 'vulnerability1')
        self.assertEqual(patch[0]['patch'], 'Dummy patch for vulnerability1')

    ## Test generate_patch method with multiple vulnerabilities
    def test_generate_patch_multiple(self):
        """
        Test that the generate_patch method works correctly with multiple vulnerabilities
        """
        patch_gen = PatchGeneration()
        patch_gen.generate_patch('vulnerability1')
        patch_gen.generate_patch('vulnerability2')
        patch = patch_gen.patches
        self.assertEqual(len(patch), 2)
        self.assertEqual(patch[0]['vulnerability'], 'vulnerability1')
        self.assertEqual(patch[0]['patch'], 'Dummy patch for vulnerability1')
        self.assertEqual(patch[1]['vulnerability'], 'vulnerability2')
        self.assertEqual(patch[1]['patch'], 'Dummy patch for vulnerability2')

    ## Test generate_patch method with empty string
    def test_generate_patch_empty(self):
        """
        Test that the generate_patch method works correctly with an empty string
        """
        patch_gen = PatchGeneration()
        patch = patch_gen.generate_patch('')
        self.assertEqual(len(patch), 1)
        self.assertEqual(patch[0]['vulnerability'], '')
        self.assertEqual(patch[0]['patch'], 'Dummy patch for ')

if __name__ == '__main__':
    unittest.main()
